Updated GMime to version 2.6.23.

+++ src/depends/gmime/gmime/gmime-message-partial.c
@@ -36,7 +36,7 @@
  * SECTION: gmime-message-partial
  * @title: GMimeMessagePartial
  * @short_description: Partial MIME parts
- * @se_also:
+ * @see_also:
  *
  * A #GMimeMessagePartial represents the message/partial MIME part.
  **/
@@ -436,7 +436,7 @@ g_mime_message_partial_split_message (GMimeMessage *message, size_t max_size, si
 	len = g_mime_stream_length (stream);
 	
 	/* optimization */
-	if (len <= max_size) {
+	if (len <= (gint64) max_size) {
 		g_object_unref (stream);
 		g_object_ref (message);
